Split dashboard class into its own file and changed tests to work accordingly
